I was bored, fixed demo so it actually works. Plus I simplified the Physics code a bit.

+++ src/engine/component/physics.h
@@ -23,7 +23,7 @@ class Physics : public Component::iComponent
 		float accstep;
 		float deaccstep;
 		
-		Interval<Uint64> timestep;
+		Interval<Uint32> timestep;
 		
 	public:
 		Physics();
